Function pallet_transaction_storage::pallet::dispatchables::store

pub fn store<T: Config>(data: Vec<u8>)

Index and store data off chain. Minimum data size is 1 bytes, maximum is MaxTransactionSize. Data will be removed after STORAGE_PERIOD blocks, unless renew is called.

Complexity

O(n*log(n)) of data size, as all data is pushed to an in-memory trie.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::store.