Trait pallet_transaction_storage::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>>;
    type Currency: Mutate<Self::AccountId> + MutateHold<Self::AccountId, Reason = Self::RuntimeHoldReason> + Balanced<Self::AccountId>;
    type RuntimeHoldReason: From<HoldReason>;
    type FeeDestination: OnUnbalanced<CreditOf<Self>>;
    type WeightInfo: WeightInfo;
    type MaxBlockTransactions: Get<u32>;
    type MaxTransactionSize: Get<u32>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>>

A dispatchable call.

type Currency: Mutate<Self::AccountId> + MutateHold<Self::AccountId, Reason = Self::RuntimeHoldReason> + Balanced<Self::AccountId>

The fungible type for this pallet.

type RuntimeHoldReason: From<HoldReason>

The overarching runtime hold reason.

type FeeDestination: OnUnbalanced<CreditOf<Self>>

Handler for the unbalanced decrease when fees are burned.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type MaxBlockTransactions: Get<u32>

Maximum number of indexed transactions in the block.

type MaxTransactionSize: Get<u32>

Maximum data set in a single transaction in bytes.

Object Safety

This trait is not object safe.