Crate pallet_transaction_storage

Transaction storage pallet. Indexes transactions and manages storage proofs.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_transaction_storage

Structs

TransactionInfo

State data for a stored transaction.

Constants

DEFAULT_MAX_BLOCK_TRANSACTIONS

DEFAULT_MAX_TRANSACTION_SIZE

Maximum bytes that can be stored in one transaction.

Type Aliases

CreditOf